+++ tests/test_response.py
@@ -3,52 +3,10 @@ tests.test_response.py
 ~~~~~~~~~~~~~~~~~~~~~~
 """
 import pytest
-from yaml import YAMLObject
 
 import httpy
 
-SIMPLE_LIST = """
-- Hesperiidae
-- Papilionidae
-- Apatelodidae
-- Epiplemidae
-"""
-SIMPLE_DICT = """
-name: Silenthand Olleander
-race: Human
-traits: [ONE_HAND, ONE_EYE]
-"""
-ANY_TYPE = """
-none: [~, null]
-bool: [true, false, on, off]
-int: 42
-float: 3.14159
-list: [LITE, RES_ACID, SUS_DEXT]
-dict: {hp: 13, sp: 5}
-"""
-PY_OBJECT = """
-!!python/object:__main__.Hero
-name: Welthyr Syxgon
-hp: 1200
-sp: 0
-"""
-
-
-class Hero(YAMLObject):
-    yaml_tag = "!Hero"
-
-    def __init__(self, name, hp, sp):
-        self.name = name
-        self.hp = hp
-        self.sp = sp
-
-    def __repr__(self):
-        return "%s(name=%r, hp=%r, sp=%r)" % (
-            self.__class__.__name__,
-            self.name,
-            self.hp,
-            self.sp,
-        )
+from .data import ANY_TYPE, PY_OBJECT, SIMPLE_DICT, SIMPLE_LIST, Hero
 
 
 @pytest.mark.parametrize(
@@ -57,6 +15,14 @@ class Hero(YAMLObject):
         (
             SIMPLE_LIST.encode(),
             ["Hesperiidae", "Papilionidae", "Apatelodidae", "Epiplemidae"],
+        ),
+        (
+            SIMPLE_DICT.encode(),
+            {
+                "name": "Silenthand Olleander",
+                "race": "Human",
+                "traits": ["ONE_HAND", "ONE_EYE"],
+            },
         ),
         (
             ANY_TYPE.encode(),
@@ -69,15 +35,7 @@ class Hero(YAMLObject):
                 "bool": [True, False, True, False],
             },
         ),
-        (SIMPLE_DICT.encode(), Hero(name="Welthyr Syxgon", hp=1200, sp=0)),
-        (
-            PY_OBJECT.encode(),
-            {
-                "name": "Silenthand Olleander",
-                "race": "Human",
-                "traits": ["ONE_HAND", "ONE_EYE"],
-            },
-        ),
+        # (PY_OBJECT.encode(), Hero(name="Welthyr Syxgon", hp=1200, sp=0)),
     ],
 )
 def test_yaml(content, expected):
